Curses, sounds, config.
- Added AbstractRandomCurse
- Added curse ambient sounds for target players.
- Rite of broiling config now uses double.

+++ src/main/java/com/favouriteless/enchanted/EnchantedConfig.java
@@ -48,7 +48,7 @@ public class EnchantedConfig {
     // Rites
     public static final ForgeConfigSpec.ConfigValue<Integer> TOTAL_ECLIPSE_COOLDOWN;
     public static final ForgeConfigSpec.ConfigValue<Integer> SKY_WRATH_COOLDOWN;
-    public static final ForgeConfigSpec.ConfigValue<Float> BROILING_BURN_CHANCE;
+    public static final ForgeConfigSpec.ConfigValue<Double> BROILING_BURN_CHANCE;
 
     static {
         BUILDER.push("Altar Options");
@@ -72,7 +72,7 @@ public class EnchantedConfig {
         BUILDER.push("Rite Options");
         TOTAL_ECLIPSE_COOLDOWN = BUILDER.comment("Global cooldown for the rite of eclipse in seconds").define("total_eclipse_cooldown", 0);
         SKY_WRATH_COOLDOWN = BUILDER.comment("Global cooldown for the rite of sky's wrath in seconds").define("sky_wrath_cooldown", 0);
-        BROILING_BURN_CHANCE = BUILDER.comment("Chance for food to be burned by the rite of broiling #default 0.3").define("broiling_burn_chance", 0.3F);
+        BROILING_BURN_CHANCE = BUILDER.comment("Chance for food to be burned by the rite of broiling #default 0.3").define("broiling_burn_chance", 0.3D);
         BUILDER.pop();
 
         BUILDER.push("Miscellaneous Options");
